Notification Schema working
