+++ hamza-server/src/api/custom/notification/add-notification/route.ts
@@ -1,0 +1,35 @@
+import type { MedusaRequest, MedusaResponse, Logger } from '@medusajs/medusa';
+import CustomerNotificationService from '../../../../services/customer-notification';
+import { readRequestBody } from '../../../../utils/request-body';
+
+export const POST = async (req: MedusaRequest, res: MedusaResponse) => {
+    const customerNotificationService: CustomerNotificationService =
+        req.scope.resolve('customerNotificationService');
+    const logger: Logger = req.scope.resolve('logger');
+
+    const { customer_id, notification_type } = readRequestBody(req.body, [
+        'customer_id',
+        'notification_type',
+    ]);
+    if (!customer_id || !notification_type) {
+        return res
+            .status(400)
+            .json({
+                message: 'customer_id and notification_type are required',
+            });
+    }
+
+    try {
+        const types = await customerNotificationService.addNotification(
+            customer_id,
+            notification_type
+        );
+
+        res.status(200).json({ types });
+    } catch (err) {
+        logger.error('Error creating notification types', err);
+        res.status(500).json({
+            error: 'Failed to create notification types',
+        });
+    }
+};
